+++ src/common/Translations.ts
@@ -1,0 +1,5 @@
+export default class Translations {
+  public static DEFAULT_TRANSLATIONS = {
+    noNoteFound: "No notes found past the current point"
+  }
+}